requirements.txt
# Core Dependencies
sentence-transformers==2.3.1
transformers==4.36.2
torch==2.1.2

# NLP & Text Processing
spacy==3.7.2
nltk==3.8.1
pypdf==3.17.4

# Graph & Community Detection
networkx==3.2.1
python-louvain==0.16
leidenalg==0.10.2
igraph==0.11.3

# Vector Storage & Search
chromadb==0.4.22
faiss-cpu==1.7.4
scikit-learn==1.4.0

# LLM Integration
langchain==0.1.4
langchain-community==0.0.16
ollama==0.1.6

# Utilities
numpy==1.26.3
pandas==2.1.4
pydantic==2.5.3
tqdm==4.66.1
pyyaml==6.0.1
